# contracts/runtime/v1/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/campaigns": {
      "post": {
        "operationId": "CampaignController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Create a text campaign draft",
        "tags": [
          "campaigns"
        ]
      },
      "get": {
        "operationId": "CampaignController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "sessionId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "List campaigns for allowlisted sessions",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}": {
      "get": {
        "operationId": "CampaignController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Read a campaign",
        "tags": [
          "campaigns"
        ]
      },
      "patch": {
        "operationId": "CampaignController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Update an editable campaign draft",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/targets": {
      "get": {
        "operationId": "CampaignController_listTargets",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignTargetListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "List selected group targets with current capability",
        "tags": [
          "campaigns"
        ]
      },
      "put": {
        "operationId": "CampaignController_replaceTargets",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceCampaignTargetsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignTargetListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Atomically replace all group targets of a campaign draft",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/preflight": {
      "post": {
        "operationId": "CampaignController_preflight",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignPreflightRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignPreflightDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Evaluate a campaign without creating a run",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/runs": {
      "post": {
        "operationId": "CampaignController_createRun",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRunDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Create an idempotent durable campaign run",
        "tags": [
          "campaigns"
        ]
      },
      "get": {
        "operationId": "CampaignController_listRuns",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "List runs of a campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaign-runs/{id}": {
      "get": {
        "operationId": "CampaignRunController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Read a durable campaign run",
        "tags": [
          "campaign-runs"
        ]
      }
    },
...
    "/api/v1/campaign-runs/{id}/pause": {
      "post": {
        "operationId": "CampaignRunController_pause",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Pause new delivery materialization for a campaign run",
        "tags": [
          "campaign-runs"
        ]
      }
    },
    "/api/v1/campaign-runs/{id}/resume": {
      "post": {
        "operationId": "CampaignRunController_resume",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Re-run preflight and resume a paused or blocked campaign run",
        "tags": [
          "campaign-runs"
        ]
      }
    },
    "/api/v1/campaign-runs/{id}/cancel": {
      "post": {
        "operationId": "CampaignRunController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Cancel pending work for a campaign run",
        "tags": [
          "campaign-runs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Automation Runtime API",
    "description": "Stable, versioned API contract consumed by WA Studio",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "runtime-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Runtime-Key"
      }
    },
    "schemas": {
...
      "GroupSendCapabilityDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ALLOWED",
              "DENIED",
              "UNKNOWN"
            ]
          },
          "reason": {
            "type": "string"
          },
          "checkedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "invalidatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "revision": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "status",
          "reason",
          "checkedAt",
          "invalidatedAt",
          "revision"
        ]
      },
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "limit",
          "offset"
        ]
      },
...
      "CreateCampaignDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "text": {
            "type": "string",
            "maxLength": 4096
          },
          "scheduleType": {
            "type": "string",
            "enum": [
              "IMMEDIATE",
              "ONCE"
            ],
            "default": "IMMEDIATE"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time",
            "description": "Required only for ONCE scheduling"
          }
        },
        "required": [
          "sessionId",
          "name",
          "text"
        ]
      },
      "CampaignDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "scheduleType": {
            "type": "string",
            "enum": [
              "IMMEDIATE",
              "ONCE"
            ]
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "PAUSED",
              "ARCHIVED"
            ]
          },
          "targetCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionId",
          "name",
          "text",
          "scheduleType",
          "scheduledAt",
          "status",
          "targetCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "CampaignListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampaignDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateCampaignDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "text": {
            "type": "string",
            "maxLength": 4096
          },
          "scheduleType": {
            "type": "string",
            "enum": [
              "IMMEDIATE",
              "ONCE"
            ]
          },
          "scheduledAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "CampaignTargetDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          },
          "groupName": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "sendCapability": {
            "$ref": "#/components/schemas/GroupSendCapabilityDto"
          }
        },
        "required": [
          "groupId",
          "groupName",
          "enabled",
          "sendCapability"
        ]
      },
      "CampaignTargetListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampaignTargetDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "ReplaceCampaignTargetsDto": {
        "type": "object",
        "properties": {
          "groupIds": {
            "maxItems": 1000,
            "example": [
              "120363000000000000@g.us"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "groupIds"
        ]
      },
      "CampaignPreflightRequestDto": {
        "type": "object",
        "properties": {
          "executionMode": {
            "type": "string",
            "enum": [
              "DRY_RUN",
              "LIVE"
            ],
            "default": "DRY_RUN"
          }
        },
        "required": [
          "executionMode"
        ]
      },
      "CampaignPreflightCheckDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PASS",
              "WARN",
              "BLOCK"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "status",
          "message"
        ]
      },
      "CampaignTargetIssueDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          },
          "groupName": {
            "type": "string"
          },
          "capability": {
            "type": "string",
            "enum": [
              "ALLOWED",
              "DENIED",
              "UNKNOWN"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "groupId",
          "groupName",
          "capability",
          "reason"
        ]
      },
      "CampaignPreflightDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PASS",
              "WARN",
              "BLOCK"
            ]
          },
          "policyVersion": {
            "type": "number"
          },
          "executionMode": {
            "type": "string",
            "enum": [
              "DRY_RUN",
              "LIVE"
            ]
          },
          "checkedAt": {
            "format": "date-time",
            "type": "string"
          },
          "totalTargets": {
            "type": "number"
          },
          "allowedTargets": {
            "type": "number"
          },
          "deniedTargets": {
            "type": "number"
          },
          "unknownTargets": {
            "type": "number"
          },
          "checks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampaignPreflightCheckDto"
            }
          },
          "targetIssues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampaignTargetIssueDto"
            }
          }
        },
        "required": [
          "status",
          "policyVersion",
          "executionMode",
          "checkedAt",
          "totalTargets",
          "allowedTargets",
          "deniedTargets",
          "unknownTargets",
          "checks",
          "targetIssues"
        ]
      },
      "CreateCampaignRunDto": {
        "type": "object",
        "properties": {
          "executionMode": {
            "type": "string",
            "enum": [
              "DRY_RUN",
              "LIVE"
            ],
            "default": "DRY_RUN"
          }
        },
        "required": [
          "executionMode"
        ]
      },
      "CampaignRunProgressDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "pending": {
            "type": "number"
          },
          "materialized": {
            "type": "number"
          },
          "processing": {
            "type": "number"
          },
          "dryRunCompleted": {
            "type": "number"
          },
          "accepted": {
            "type": "number"
          },
          "sent": {
            "type": "number"
          },
          "delivered": {
            "type": "number"
          },
          "read": {
            "type": "number"
          },
          "failed": {
            "type": "number"
          },
          "unknown": {
            "type": "number"
          },
          "blocked": {
            "type": "number"
          },
          "cancelled": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "pending",
          "materialized",
          "processing",
          "dryRunCompleted",
          "accepted",
          "sent",
          "delivered",
          "read",
          "failed",
          "unknown",
          "blocked",
          "cancelled"
        ]
      },
      "CampaignRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "executionMode": {
            "type": "string",
            "enum": [
              "DRY_RUN",
              "LIVE"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PREPARING",
              "BLOCKED",
              "SCHEDULED",
              "RUNNING",
              "PAUSED",
              "COMPLETED",
              "PARTIAL_FAILED",
              "CANCELLED",
              "FAILED"
            ]
          },
          "statusReason": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string"
          },
          "preflight": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CampaignPreflightDto"
              }
            ]
          },
          "totalTargets": {
            "type": "number"
          },
          "progress": {
            "$ref": "#/components/schemas/CampaignRunProgressDto"
          },
          "scheduledAt": {
            "format": "date-time",
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "campaignId",
          "sessionId",
          "executionMode",
          "status",
          "statusReason",
          "text",
          "preflight",
          "totalTargets",
          "progress",
          "scheduledAt",
          "startedAt",
          "completedAt",
          "createdAt"
        ]
      },
      "CampaignRunListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampaignRunDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}